Rename project from swver to swinfo
This is because the former gives the impression that is all about
VERSION even it output a software name and its release date.

I believe that the latter is much better, as to have a wide scope to
improve regarding the project as I have a lot of ideas to implement.